requirements.txt
aiohttp==3.11.18
PyYAML==6.0.2
feedparser==6.0.11
python-dotenv==1.0.1
